+++ api/nodes.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"log"
 	"net/http"
-	"time"
 
 	"github.com/simpleiot/simpleiot/data"
 	"github.com/simpleiot/simpleiot/db/genji"
@@ -183,33 +182,6 @@ func (h *Nodes) ServeHTTP(res http.ResponseWriter, req *http.Request) {
 		default:
 			http.Error(res, "invalid method", http.StatusMethodNotAllowed)
 		}
-
-	case "cmd":
-		if req.Method == http.MethodPost {
-			// Post is typically done by UI
-			if !validUser {
-				http.Error(res, "Unauthorized", http.StatusUnauthorized)
-				return
-			}
-
-			h.processCmd(res, req, id)
-			return
-		} else if req.Method == http.MethodGet {
-			cmd, err := h.db.NodeGetCmd(id)
-			if err != nil {
-				http.Error(res, err.Error(), http.StatusInternalServerError)
-			}
-
-			// id is not required
-			cmd.ID = ""
-
-			en := json.NewEncoder(res)
-			en.Encode(cmd)
-			return
-		} else {
-			http.Error(res, "only POST/GET allowed", http.StatusMethodNotAllowed)
-			return
-		}
 	}
 }
 
@@ -232,55 +204,6 @@ func (h *Nodes) insertNode(res http.ResponseWriter, req *http.Request) {
 	}
 
 	encode(res, data.StandardResponse{Success: true, ID: id})
-}
-
-func (h *Nodes) processCmd(res http.ResponseWriter, req *http.Request, id string) {
-	decoder := json.NewDecoder(req.Body)
-	var cmd data.NodeCmd
-	err := decoder.Decode(&cmd)
-	if err != nil {
-		http.Error(res, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	// set ID in case it is not set in API call
-	cmd.ID = id
-
-	// set cmd in DB for legacy nodes that still fetch over http
-	err = h.db.NodeSetCmd(cmd)
-	if err != nil {
-		http.Error(res, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	// TODO how to support old nodes still fetching commands via http
-	// perhaps check if node is connected via NATs
-	err = nats.SendCmd(h.nh.Nc, cmd, time.Second*10)
-	if err != nil {
-		log.Printf("Error sending command (%v) to node: ", err)
-		// don't return HTTP error for now as some units still fetch over http
-		//http.Error(res, "Error sending command to node", http.StatusInternalServerError)
-		//return
-	} else {
-		err = h.db.NodeDeleteCmd(cmd.ID)
-		if err != nil {
-			log.Printf("Error deleting command for node %v: %v", id, err)
-		}
-	}
-
-	// process updates that are now pushed
-	if cmd.Cmd == data.CmdUpdateApp {
-		log.Printf("Sending %v to node %v\n", cmd.Detail, cmd.ID)
-		err := h.nh.StartUpdate(cmd.ID, cmd.Detail)
-		if err != nil {
-			log.Println("Error starting app update: ", err)
-			http.Error(res, "error starting update", http.StatusInternalServerError)
-			return
-		}
-	}
-
-	en := json.NewEncoder(res)
-	en.Encode(data.StandardResponse{Success: true, ID: id})
 }
 
 func (h *Nodes) processPoints(res http.ResponseWriter, req *http.Request, id string) {