+++ node/node.go
@@ -59,19 +59,6 @@ func (m *Manager) Run() {
 					log.Println("Error updating node state: ", err)
 				}
 			}
-
-			for _, ruleID := range node.Rules {
-				rule, err := m.db.RuleByID(ruleID)
-				if err != nil {
-					log.Printf("Error finding rule %v: %v\n", ruleID, err)
-					continue
-				}
-
-				err = m.runRule(&node, &rule)
-				if err != nil {
-					log.Println("Error running rule: ", ruleID)
-				}
-			}
 		}
 
 		time.Sleep(1 * time.Second)
@@ -88,40 +75,6 @@ func uniqueUsers(users []data.User) []data.User {
 	}
 
 	return ret
-}
-
-func (m *Manager) runRule(node *data.Node, rule *data.Rule) error {
-	if node.State() != data.PointValueSysStateOnline {
-		// only run rules if node is in online state
-		return nil
-	}
-
-	active := rule.IsActive(node.Points)
-	if active != rule.State.Active {
-		state := data.RuleState{Active: active}
-		if active {
-			// process actions
-			if !rule.State.Active && rule.Config.Repeat == 0 {
-				for _, a := range rule.Config.Actions {
-					if a.Type == data.ActionTypeNotify {
-						err := m.notify(node, rule.Config.Description, a.Template, node.Groups)
-						if err != nil {
-							log.Println("Error notifying: ", err)
-						}
-					}
-				}
-				state.LastAction = time.Now()
-			}
-		}
-
-		// store updated state in DB
-		err := m.db.RuleUpdateState(rule.ID, state)
-		if err != nil {
-			log.Println("Error updating rule state: ", err)
-		}
-	}
-
-	return nil
 }
 
 func (m *Manager) notify(node *data.Node, ruleDesc, msgTemplate string, groups []string) error {