+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For more details or to discuss releases, please visit the
 ## Next
 
 - UI: change node min/max button to edit/close
+- Modbus: suppress TCP conn/disc messages at debug level 0
 
 ## [[0.0.21] - 2021-03-17](https://github.com/simpleiot/simpleiot/releases/tag/v0.0.21)