+++ CHANGELOG.md
@@ -13,6 +13,8 @@ For more details or to discuss releases, please visit the
 
 - UI: change node min/max button to edit/close
 - Modbus: suppress TCP conn/disc messages at debug level 0
+- siot: add cmdline option (-logNats) to trace all node points. This can be run
+  in parallel to the siot application to trace points flowing through the system
 
 ## [[0.0.21] - 2021-03-17](https://github.com/simpleiot/simpleiot/releases/tag/v0.0.21)