+++ CHANGELOG.md
@@ -11,7 +11,12 @@ For more details or to discuss releases, please visit the
 
 ## Next
 
+- UI: change node min/max button to edit/close
+- Modbus: suppress TCP conn/disc messages at debug level 0
+- siot: add cmdline option (-logNats) to trace all node points. This can be run
+  in parallel to the siot application to trace points flowing through the system
 - genji db: update to v0.11.0 release
+- rules: can now write rules that set nodes based on other nodes
 
 ## [[0.0.21] - 2021-03-17](https://github.com/simpleiot/simpleiot/releases/tag/v0.0.21)